Revert "Fix type hint"
This reverts commit f684f0fcd9e177b59e5efa4dd66c6148cef640cd.

+++ mobility_on_demand/model/dispatch.py
@@ -35,7 +35,7 @@ class Dispatcher:
 
 
 class ScoredCandidate:
-    def __init__(self, candidate: DispatchCandidate, score: float):
+    def __init__(self, candidate: DispatchCandidate, score: int):
         self.candidate = candidate
         self.score = score
